requirements.txt
PyPDF2>=3.0.0
Pillow>=9.0.0
moviepy>=1.0.3
pdf2image>=1.16.0
numpy>=1.20.0
opencv-python>=4.5.0
pytesseract>=0.3.9
fpdf>=1.7.2
tqdm>=4.62.0
